Atualização de plugins

+++ wp-content/plugins/advanced-database-cleaner/includes/classes/general-cleanup/class-adbc-abstract-cleanup-handler.php
@@ -910,6 +910,12 @@ abstract class ADBC_Abstract_Cleanup_Handler implements ADBC_Cleanup_Type_Handle
 
 			ADBC_Sites::instance()->switch_to_blog_id( $site['id'] );
 
+			// Make sure the tables exists in this blog
+			if ( ! ADBC_Tables::is_table_exists( $this->table() ) ) {
+				ADBC_Sites::instance()->restore_blog();
+				continue; // nothing to do on this blog
+			}
+
 			// Wrap in a derived table to dodge MySQL error 1093
 			$sql = "
 				DELETE FROM {$this->table()}
